Long int to unsigned long int

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,18 +8,18 @@
  */
 int main(void)
 {
-	long long int a, b, c, i;
+	unsigned long int a, b, c, i;
 
 	a = 0;
 	b = 1;
 	c = a + b;
-	printf("%ld, ", c);
+	printf("%lu, ", c);
 	for (i = 2; i <= 98; i++)
 	{
 		a = b;
 		b = c;
 		c = a + b;
-		printf("%ld", c);
+		printf("%lu", c);
 		if (i == 98)
 			continue;
 		putchar(',');
